cleaning cip2 tests
